app/requirements.txt
beautifulsoup4==4.11.1
bs4==0.0.1
cachetools==4.2.2
certifi==2022.9.24
charset-normalizer==2.1.1
google-api-core==2.10.1
google-api-python-client==2.63.0
google-auth==2.12.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.56.4
httplib2==0.20.4
idna==3.4
oauth2client==4.1.3
protobuf==4.21.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
python-dotenv==0.21.0
requests==2.28.1
rsa==4.9
six==1.16.0
soupsieve==2.3.2.post1
uritemplate==4.1.1
urllib3==1.26.12
lxml==4.9.1
psycopg2-binary==2.9.3
python-telegram-bot==13.14